pyproject.toml
[tool.poetry]
name = "bra-interface"
version = "0.1.0"
description = ""
authors = ["Emeric <emeric.dynomant@gmail.com>"]
license = "GPL-3"

packages = [
    { include = "bra_interface" }
]

[tool.poetry.dependencies]
python = "^3.9"
PyMySQL = "^1.0.2"
python-dotenv = "^0.19.2"
pandas = "^1.4.1"
folium = "^0.12.1"
wheel = "^0.37.1"
Shapely = "^1.8.1"
geopandas = "^0.10.2"
libpysal = "^4.6.2"
matplotlib = "^3.5.1"
mapclassify = "^2.4.3"
PyQt5 = "^5.15.6"
fastparquet = "^0.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
